moved validation to download component
