+++ assets/training/model_management/components/download_model/spec.yaml
@@ -16,6 +16,7 @@ command: >
   python run_model_download.py
   --model-source ${{inputs.model_source}}
   --model-id '${{inputs.model_id}}'
+  --update-existing-model ${{inputs.update_existing_model}}
   --model-download-metadata ${{outputs.model_download_metadata}}
   --model-output-dir ${{outputs.model_output}}
 
@@ -36,6 +37,12 @@ inputs:
   validation_info:
     type: uri_file
     description: Path to the validation info file
+    optional: true
+
+  update_existing_model:
+    type: boolean
+    default: false
+    description: If set to true, will update the existing model. If set to false, will create a new model.
     optional: true
 
 outputs:
